scanner_agent_311/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
langchain==0.3.15
python-dotenv
ipython
pillow
graphviz